Good enough for now

+++ Assets/Code/Scripts/IDamagable.cs
@@ -1,8 +1,9 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
 public interface IDamagable 
 {
-    public void TakeDamage(float _damage);
+    public void TakeDamage(int _damage);
 }
